Version 3.0
-Allows users to download and compile maximum videos from youtube in many formats.
-Sublimed the interface a little 
-This version also comes with a separate merger program for those videos that weren't merged in the downloader program
\------------------Tips--------------\ 
For a better experience please use download ids 399 or 398 or simply stick to 60fps videos.
Although the merger program is able to merge videos up to 4k quality it's ill-advised to do so [It lags although more testing hasn't be done].
Save the files and folders in a single directory (i.e set the save paths the same )to avoid the error no files found. 

Special thanks to tkhalim14 for contributing to the project.